[Security Solution] unwire dead sourcerer redux slice from the store and delete slice files
