+++ x-pack/solutions/security/plugins/security_solution/public/common/store/types.ts
@@ -14,7 +14,6 @@ import type { AppAction } from './actions';
 import type { Immutable } from '../../../common/endpoint/types';
 import type { AppState } from './app/reducer';
 import type { InputsState } from './inputs/reducer';
-import type { SourcererState } from '../../sourcerer/store/reducer';
 import type { HostsPluginState } from '../../explore/hosts/store';
 import type { DragAndDropState } from './drag_and_drop/reducer';
 import type { TimelinePluginState } from '../../timelines/store';
@@ -36,7 +35,6 @@ export type State = HostsPluginState &
     app: AppState;
     dragAndDrop: DragAndDropState;
     inputs: InputsState;
-    sourcerer: SourcererState;
     globalUrlParam: GlobalUrlParam;
     discover: SecuritySolutionDiscoverState;
   } & DataTableState &
