--- x-pack/solutions/security/plugins/security_solution/public/sourcerer/store/helpers.ts
@@ -1,133 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-import { isEmpty } from 'lodash';
-import { PageScope } from '../../data_view_manager/constants';
-import type { SourcererDataView, SourcererModel, SourcererScopeById } from './model';
-import type { sourcererModel } from '.';
-import type { SelectedDataViewPayload } from './actions';
-import { ensurePatternFormat, sortWithExcludesAtEnd } from '../../../common/utils/sourcerer';
-
-const getPatternListFromScope = (
-  scope: PageScope,
-  patternList: string[],
-  signalIndexName: string | null
-) => {
-  // when our SIEM data view is set, here are the defaults
-  switch (scope) {
-    case PageScope.default:
-    case PageScope.explore:
-      return sortWithExcludesAtEnd(patternList.filter((index) => index !== signalIndexName));
-    case PageScope.alerts:
-      // set to signalIndexName whether or not it exists yet in the patternList
-      return signalIndexName != null ? [signalIndexName] : [];
-    case PageScope.attacks:
-      return sortWithExcludesAtEnd(patternList);
-    case PageScope.timeline:
-      return sortWithExcludesAtEnd(patternList);
-    case PageScope.analyzer:
-      return sortWithExcludesAtEnd(patternList);
-  }
-};
-
-export const getScopePatternListSelection = (
-  theDataView: SourcererDataView | undefined,
-  sourcererScope: PageScope,
-  signalIndexName: SourcererModel['signalIndexName'],
-  isDefaultDataView: boolean
-): string[] => {
-  const patternList: string[] =
-    theDataView != null && theDataView.id !== null ? theDataView.patternList : [];
-
-  if (!isDefaultDataView) {
-    return sortWithExcludesAtEnd(patternList);
-  }
-
-  return getPatternListFromScope(sourcererScope, patternList, signalIndexName);
-};
-
-export const validateSelectedPatterns = (
-  state: SourcererModel,
-  payload: SelectedDataViewPayload,
-  shouldValidateSelectedPatterns: boolean
-): Partial<SourcererScopeById> => {
-  const { id, ...rest } = payload;
-  const dataView = state.kibanaDataViews.find((p) => p.id === rest.selectedDataViewId);
-  // dedupe because these could come from a silly url or pre 8.0 timeline
-  const dedupePatterns = ensurePatternFormat(rest.selectedPatterns);
-  let selectedPatterns =
-    // shouldValidateSelectedPatterns is false when upgrading from
-    // legacy pre-8.0 timeline index patterns to data view.
-    shouldValidateSelectedPatterns &&
-    dataView != null &&
-    // don't validate when the data view has not been initialized (default is initialized already always)
-    dataView.id !== state.defaultDataView.id &&
-    dataView.patternList.length > 0
-      ? dedupePatterns.filter(
-          (pattern) =>
-            (dataView != null && dataView.patternList.includes(pattern)) ||
-            // this is a hack, but sometimes signal index is deleted and is getting regenerated. it gets set before it is put in the dataView
-            state.signalIndexName == null ||
-            state.signalIndexName === pattern
-        )
-      : // don't remove non-existing patterns, they were saved in the first place in timeline
-        // but removed from the security data view
-        // or its a legacy pre-8.0 timeline
-        dedupePatterns;
-  const signalIndexName = state.signalIndexName;
-  selectedPatterns = getPatternListFromScope(id, selectedPatterns, signalIndexName);
-
-  return {
-    [id]: {
-      ...state.sourcererScopes[id],
-      ...rest,
-      selectedDataViewId: dataView?.id ?? null,
-      selectedPatterns,
-      // if in timeline, allow for empty in case pattern was deleted
-      // need flow for this
-      ...(isEmpty(selectedPatterns) && id !== PageScope.timeline
-        ? {
-            selectedPatterns: getScopePatternListSelection(
-              dataView ?? state.defaultDataView,
-              id,
-              state.signalIndexName,
-              (dataView ?? state.defaultDataView).id === state.defaultDataView.id
-            ),
-          }
-        : {}),
-      loading: false,
-    },
-  };
-};
-
-interface CheckIfIndicesExistParams {
-  patternList: sourcererModel.SourcererDataView['patternList'];
-  scopeId: PageScope;
-  signalIndexName: string | null;
-  isDefaultDataViewSelected: boolean;
-}
-
-export const checkIfIndicesExist = ({
-  patternList,
-  scopeId,
-  signalIndexName,
-  isDefaultDataViewSelected,
-}: CheckIfIndicesExistParams) => {
-  if (scopeId === PageScope.alerts) {
-    return patternList.includes(`${signalIndexName}`);
-  }
-
-  if (scopeId === PageScope.default) {
-    if (isDefaultDataViewSelected) {
-      return patternList.filter((i) => i !== signalIndexName).length > 0;
-    }
-
-    return patternList.length > 0;
-  }
-
-  return patternList.length > 0;
-};
