--- x-pack/solutions/security/plugins/security_solution/public/sourcerer/store/index.ts
@@ -1,12 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-import * as sourcererActions from './actions';
-import * as sourcererModel from './model';
-
-export { sourcererActions, sourcererModel };
-export * from './reducer';
