+++ x-pack/solutions/security/plugins/security_solution/public/timelines/hooks/use_create_timeline.test.tsx
@@ -15,7 +15,7 @@ import { timelineActions } from '../store';
 import { inputsActions } from '../../common/store/inputs';
 import { appActions } from '../../common/store/app';
 import { InputsModelId } from '../../common/store/inputs/constants';
-import { mockGlobalState, TestProviders } from '../../common/mock';
+import { TestProviders } from '../../common/mock';
 import { defaultUdtHeaders } from '../components/timeline/body/column_headers/default_headers';
 import { PageScope } from '../../data_view_manager/constants';
 import { useSecurityDefaultPatterns } from '../../data_view_manager/hooks/use_security_default_patterns';
@@ -36,6 +36,22 @@ jest.mock('../../common/lib/kibana');
 jest.mock('../../data_view_manager/hooks/use_security_default_patterns');
 jest.mock('../../data_view_manager/hooks/use_select_data_view');
 
+const mockDefaultPatterns = {
+  id: 'security-solution',
+  indexPatterns: [
+    '.siem-signals-spacename',
+    'apm-*-transaction*',
+    'auditbeat-*',
+    'endgame-*',
+    'filebeat-*',
+    'logs-*',
+    'packetbeat-*',
+    'traces-apm*',
+    'winlogbeat-*',
+    '-*elastic-cloud-logs-*',
+  ],
+};
+
 describe('useCreateTimeline', () => {
   const resetDiscoverAppState = jest.fn().mockResolvedValue({});
   const selectDataView = jest.fn();
@@ -44,21 +60,7 @@ describe('useCreateTimeline', () => {
     jest.clearAllMocks();
     (useDiscoverInTimelineContext as jest.Mock).mockReturnValue({ resetDiscoverAppState });
     (useSelectDataView as jest.Mock).mockReturnValue(selectDataView);
-    (useSecurityDefaultPatterns as jest.Mock).mockReturnValue({
-      id: 'security-solution',
-      indexPatterns: [
-        '.siem-signals-spacename',
-        'apm-*-transaction*',
-        'auditbeat-*',
-        'endgame-*',
-        'filebeat-*',
-        'logs-*',
-        'packetbeat-*',
-        'traces-apm*',
-        'winlogbeat-*',
-        '-*elastic-cloud-logs-*',
-      ],
-    });
+    (useSecurityDefaultPatterns as jest.Mock).mockReturnValue(mockDefaultPatterns);
   });
 
   it('should return a function', () => {
@@ -96,25 +98,17 @@ describe('useCreateTimeline', () => {
     expect(createTimeline.mock.calls[0][0].id).toEqual(TimelineId.test);
     expect(createTimeline.mock.calls[0][0].timelineType).toEqual(TimelineTypeEnum.default);
     expect(createTimeline.mock.calls[0][0].columns).toEqual(defaultUdtHeaders);
-    expect(createTimeline.mock.calls[0][0].dataViewId).toEqual(
-      mockGlobalState.sourcerer.defaultDataView.id
-    );
+    expect(createTimeline.mock.calls[0][0].dataViewId).toEqual(mockDefaultPatterns.id);
     expect(createTimeline.mock.calls[0][0].indexNames).toEqual(
-      expect.arrayContaining(
-        mockGlobalState.sourcerer.sourcererScopes[PageScope.timeline].selectedPatterns
-      )
+      expect.arrayContaining(mockDefaultPatterns.indexPatterns)
     );
     expect(createTimeline.mock.calls[0][0].show).toEqual(true);
     expect(createTimeline.mock.calls[0][0].updated).toEqual(undefined);
     expect(createTimeline.mock.calls[0][0].excludedRowRendererIds).toHaveLength(RowRendererCount);
     expect(selectDataView.mock.calls[0][0].scope).toEqual(PageScope.timeline);
-    expect(selectDataView.mock.calls[0][0].id).toEqual(
-      mockGlobalState.sourcerer.defaultDataView.id
-    );
+    expect(selectDataView.mock.calls[0][0].id).toEqual(mockDefaultPatterns.id);
     expect(selectDataView.mock.calls[0][0].fallbackPatterns).toEqual(
-      expect.arrayContaining(
-        mockGlobalState.sourcerer.sourcererScopes[PageScope.timeline].selectedPatterns
-      )
+      expect.arrayContaining(mockDefaultPatterns.indexPatterns)
     );
     expect(addLinkTo).toHaveBeenCalledWith([InputsModelId.global, InputsModelId.timeline]);
     expect(addNotes).toHaveBeenCalledWith({ notes: [] });
